+++ app/src/main/java/com/github/andreyasadchy/xtra/repository/TwitchEndpoints.kt
@@ -1,0 +1,5 @@
+package com.github.andreyasadchy.xtra.repository
+
+internal object TwitchEndpoints {
+    const val GRAPHQL = "https://gql.twitch.tv/gql"
+}